fix(restore/swarm): init primary endpoint after admin user is created
